Split ci-notify report code by purpose
